requirements.txt
transformers
sentence-transformers
streamlit
langchain
langchain-community
faiss-cpu
trafilatura
python-dotenv
